Always reply back with URI instead of an error message
Presently, when an exception is raised with a message, for example:

   ```
   wamp_error(WAMP_ERROR_NOT_AUTHORIZED, "not authorised)
   ```

depending on how the exception is caught, the second part of the
exception (e.g. "no authorised above) is used as the error URI. This is
in violation of the WAMP specification.

This commit fixes this by placing the error description in the `Details`
part of the WAMP reply, under the "reason" key.