Don't change error URI of thrown `wamp_error` if authorisation fails
The logic for this is that if a wampcc::wamp_error has been raised by
the authorising function, then it should passed as is, because
presumably it's an error message that has a correct WAMP error URI that
makes sense to the user and may given information on why the
weatherization failed.

Other exceptions should however be re-thrown as wamp_err.