Fix getting thread ID on macOS systems
Currently, on non-Windows systems we do a `syscall(SYS_gettid)` to get
the thread ID. However, this is Linux-only and is broken macos. On macOS
the syscal to get the thread ID is `SYS_thread_selfid`.

In any case, making system calls has been deprecated in the latest
version of macOS. This commit fixes it by using the Apple specific
`pthread_threadid_np()` function instead. Note that this function
returns the same thread ID as the system call.

Technically we should also change the return type of the
function (i.e. `DWORD` for Windows, `pid_t` for Linux, etc.), but I
don't think that's necessary yet as thread IDs are not that large.