Export custom `Find<Package>.cmake` recipes when exporting config
On most systems Jansson and UV don't come with a
`Find<PackageName>.cmake` or `<PackageName>Config.cmake` files, so in
wampcc we are using custom find module recipes to find Jansson and UV.

This commit causes these custom find recipes to also be exported as part
of an install. This way UV and Jansson are automatically found when a
user does `find_package(wampcc)` in a CMake recipe.