Link wampcc against pthreads on Linux
This is because in some versions of libstdc++ (at least on 10.2.1),
`std::condition_variable::wait_for()` is implemented using
`pthread_cond_clockwait()`, and so one needs to link against pthreads.

Although this linking was always required, in previous version of
libstdc++ there was a bug: if pthread was not linked against then there
was no error message but `std::condition_variable::wait_for()` would
immediately return.